requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
openai==1.70.0
pandas==2.2.3
mlflow==2.21.2
llmevalkit
sentence-transformers
python-dotenv
groq
